refactor(api): git merge_requests 라우트를 전용 모듈로 추출
9개 라우트를 api/git/merge_requests.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 63건 동일, 가림 쌍 0건.

원본의 "Internal Merge Request Endpoints" 섹션이 연속 구간이라 통째로 옮겼다.
_legacy.py 에서 MergeRequest 계열 모델 5종과 _shared 의존 2종
(_get_db_session·get_mr_service_for_project)이 미사용이 되어 제거했다.
can_merge_to_branch 는 merge 섹션(Task 9)이 아직 쓰므로 양쪽에 남는다.

이 추출로 _legacy.py 가 528줄이 되어 처음으로 800줄 한도 아래로 내려왔다.
